Mainly writing argument parser
This commit contains some tests with regards to sourcing the profile before running commands, that way everything is available when running the command. The biggest part of this commit is the beginning of the argument parser. So far this works in a very basic fashion, but it could be quite a bit of work to make it relatively robust